+++ levenshtein/models.py
@@ -7,7 +7,7 @@ class Levenshtein(models.Model):
     text_1 = models.TextField()
     text_2 = models.TextField()
     levenshtein_value = models.IntegerField(null=True, blank=True)
-    levenshtein_radio_info = models.FloatField(null=True, blank=True)
+    levenshtein_ratio = models.FloatField(null=True, blank=True)
 
     # Timestamps
     created = models.DateTimeField(auto_now_add=True)
